# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Paradox Connect API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/hostels": {
      "get": {
        "tags": [
          "Hostels"
        ],
        "summary": "List Hostels",
        "operationId": "list_hostels_hostels_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Hostels"
        ],
        "summary": "Create Hostel",
        "operationId": "create_hostel_hostels_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostelCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/hostels/{hostel_id}/team": {
      "post": {
        "tags": [
          "Hostels"
        ],
        "summary": "Assign Hostel Team",
        "operationId": "assign_hostel_team_hostels__hostel_id__team_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "hostel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hostel Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HostelAssignTeamRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/hostels/{hostel_id}/team/{team_user_id}/toggle_scan": {
      "put": {
        "tags": [
          "Hostels"
        ],
        "summary": "Toggle Hostel Scan",
        "operationId": "toggle_hostel_scan_hostels__hostel_id__team__team_user_id__toggle_scan_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "hostel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hostel Id"
            }
          },
          {
            "name": "team_user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team User Id"
            }
          },
          {
            "name": "logging",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean",
              "title": "Logging"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/hostels/allocate": {
      "post": {
        "tags": [
          "Hostels"
        ],
        "summary": "Allocate Hostels",
        "operationId": "allocate_hostels_hostels_allocate_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/hostels/my_hostel": {
      "get": {
        "tags": [
          "Hostels"
        ],
        "summary": "My Hostel",
        "operationId": "my_hostel_hostels_my_hostel_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/hostels/{hostel_id}/scan": {
      "post": {
        "tags": [
          "Hostels"
        ],
        "summary": "Scan Hostel",
        "operationId": "scan_hostel_hostels__hostel_id__scan_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "hostel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hostel Id"
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Action"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanQRRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/hostels/{hostel_id}/statistics": {
      "get": {
        "tags": [
          "Hostels"
        ],
        "summary": "Hostel Statistics",
        "operationId": "hostel_statistics_hostels__hostel_id__statistics_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "hostel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Hostel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HostelAssignTeamRequest": {
        "properties": {
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "type": "object",
        "required": [
          "role"
        ],
        "title": "HostelAssignTeamRequest"
      },
      "HostelCreateRequest": {
        "properties": {
          "hostel_id": {
            "type": "string",
            "title": "Hostel Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "gender": {
            "type": "string",
            "title": "Gender"
          },
          "coordinator": {
            "additionalProperties": true,
            "type": "object",
            "title": "Coordinator"
          }
        },
        "type": "object",
        "required": [
          "hostel_id",
          "name",
          "capacity",
          "gender",
          "coordinator"
        ],
        "title": "HostelCreateRequest"
      },
...
      "ScanQRRequest": {
        "properties": {
          "participant_id": {
            "type": "string",
            "title": "Participant Id"
          },
          "data": {
            "type": "string",
            "title": "Data"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "participant_id",
          "data",
          "timestamp"
        ],
        "title": "ScanQRRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}